Move AnalyticsManager to its own package
