+++ app/src/main/java/com/connfa/ui/activity/AboutDetailsActivity.java
@@ -10,9 +10,9 @@ import android.webkit.WebViewClient;
 import android.widget.ScrollView;
 
 import com.connfa.R;
+import com.connfa.analytics.AnalyticsManager;
 import com.connfa.model.Model;
 import com.connfa.model.UpdatesManager;
-import com.connfa.utils.AnalyticsManager;
 import com.connfa.utils.WebviewUtils;
 import com.google.android.gms.analytics.GoogleAnalytics;
