+++ app/src/main/java/com/connfa/ui/activity/SplashActivity.java
@@ -8,12 +8,12 @@ import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 
 import com.connfa.R;
+import com.connfa.analytics.AnalyticsManager;
 import com.connfa.model.Model;
 import com.connfa.model.PreferencesManager;
 import com.connfa.model.UpdateCallback;
 import com.connfa.model.UpdatesManager;
 import com.connfa.ui.dialog.NoConnectionDialog;
-import com.connfa.utils.AnalyticsManager;
 import com.connfa.utils.NetworkUtils;
 import com.ls.util.L;
